refactor: move calculation functions
